+++ simconnect-sdk/src/domain/client_event.rs
@@ -1,6 +1,10 @@
 use std::os::raw::c_char;
 
-use crate::{bindings, SimConnectError};
+use crate::{
+    bindings,
+    helpers::{bcd16_to_u16, u16_to_bcd16},
+    SimConnectError,
+};
 
 // System Events start from 0 so we have to stagger the values to avoid collisions.
 pub(crate) const CLIENT_EVENT_DISCRIMINANT_START: u32 = 1024;
@@ -44,6 +48,24 @@ pub enum ClientEventRequest {
     AxisRightBrakeSet,
     /// Toggles parking brake on/off.
     ParkingBrakes,
+    // Comm Radios
+    Com1RadioStbySet,
+    Com2RadioStbySet,
+    Com3RadioStbySet,
+    Com1RadioSwap,
+    Com2RadioSwap,
+    Com3RadioSwap,
+    // Nav Radios
+    Nav1RadioStbySet,
+    Nav2RadioStbySet,
+    Nav3RadioStbySet,
+    Nav4RadioStbySet,
+    Nav1RadioSwap,
+    Nav2RadioSwap,
+    Nav3RadioSwap,
+    Nav4RadioSwap,
+    // Transponder
+    TransponderSet,
 }
 
 impl ClientEventRequest {
@@ -63,6 +85,24 @@ impl ClientEventRequest {
             Self::AxisLeftBrakeSet => "AXIS_LEFT_BRAKE_SET\0".as_ptr() as *const c_char,
             Self::AxisRightBrakeSet => "AXIS_RIGHT_BRAKE_SET\0".as_ptr() as *const c_char,
             Self::ParkingBrakes => "PARKING_BRAKES\0".as_ptr() as *const c_char,
+            // Comm Radios
+            Self::Com1RadioStbySet => "COM_STBY_RADIO_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Com2RadioStbySet => "COM2_STBY_RADIO_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Com3RadioStbySet => "COM3_STBY_RADIO_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Com1RadioSwap => "COM_RADIO_SWAP\0".as_ptr() as *const c_char,
+            Self::Com2RadioSwap => "COM2_RADIO_SWAP\0".as_ptr() as *const c_char,
+            Self::Com3RadioSwap => "COM3_RADIO_SWAP\0".as_ptr() as *const c_char,
+            // Nav Radios
+            Self::Nav1RadioStbySet => "NAV1_STBY_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Nav2RadioStbySet => "NAV2_STBY_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Nav3RadioStbySet => "NAV3_STBY_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Nav4RadioStbySet => "NAV4_STBY_SET_HZ\0".as_ptr() as *const c_char,
+            Self::Nav1RadioSwap => "NAV1_RADIO_SWAP\0".as_ptr() as *const c_char,
+            Self::Nav2RadioSwap => "NAV2_RADIO_SWAP\0".as_ptr() as *const c_char,
+            Self::Nav3RadioSwap => "NAV3_RADIO_SWAP\0".as_ptr() as *const c_char,
+            Self::Nav4RadioSwap => "NAV4_RADIO_SWAP\0".as_ptr() as *const c_char,
+            // Transponder
+            Self::TransponderSet => "XPNDR_SET\0".as_ptr() as *const c_char,
         }
     }
 }
@@ -71,7 +111,7 @@ impl ClientEventRequest {
 ///
 /// Defined by <https://www.prepar3d.com/SDKv5/sdk/references/variables/event_ids.html>.
 /// Extended by <https://docs.flightsimulator.com/html/Programming_Tools/Event_IDs/Event_IDs.htm>.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, PartialEq)]
 #[non_exhaustive]
 pub enum ClientEvent {
     // ---------------
@@ -126,6 +166,60 @@ pub enum ClientEvent {
     },
     /// Toggles parking brake on/off.
     ParkingBrakes,
+    /// Aircraft Radio Navigation Events
+    /// https://docs.flightsimulator.com/html/Programming_Tools/Event_IDs/Aircraft_Radio_Navigation_Events.htm
+    /// Set COM 1 radio frequency.
+    Com1RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    /// Set COM 2 radio frequency.
+    Com2RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    /// Set COM 2 radio frequency.
+    Com3RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    /// Swap COM 1 standby to active frequency.
+    Com1RadioSwap,
+    /// Swap COM 2 standby to active frequency.
+    Com2RadioSwap,
+    /// Swap COM 3 standby to active frequency.
+    Com3RadioSwap,
+    // Nav Radios
+    Nav1RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    Nav2RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    Nav3RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    Nav4RadioStbySet {
+        /// Value is desired frequency eg. 127.65
+        value: f32,
+    },
+    /// Swap NAV 1 standby to active frequency.
+    Nav1RadioSwap,
+    /// Swap NAV 2 standby to active frequency.
+    Nav2RadioSwap,
+    /// Swap NAV 3 standby to active frequency.
+    Nav3RadioSwap,
+    /// Swap NAV 4 standby to active frequency.
+    Nav4RadioSwap,
+    /// Sets the Transponder value.
+    TransponderSet {
+        /// Transponder value eg. 0174
+        /// Largest individual digit is 7, transponder values will overflow.
+        value: u16,
+    },
 }
 
 impl TryFrom<&bindings::SIMCONNECT_RECV_EVENT> for ClientEvent {
@@ -164,31 +258,121 @@ impl TryFrom<&bindings::SIMCONNECT_RECV_EVENT> for ClientEvent {
                 value: event.dwData as i32,
             }),
             ClientEventRequest::ParkingBrakes => Ok(Self::ParkingBrakes),
+            // Comm Radios
+            ClientEventRequest::Com1RadioStbySet => Ok(Self::Com1RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Com2RadioStbySet => Ok(Self::Com2RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Com3RadioStbySet => Ok(Self::Com3RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Com1RadioSwap => Ok(Self::Com1RadioSwap),
+            ClientEventRequest::Com2RadioSwap => Ok(Self::Com2RadioSwap),
+            ClientEventRequest::Com3RadioSwap => Ok(Self::Com3RadioSwap),
+            // Nav Radios
+            ClientEventRequest::Nav1RadioStbySet => Ok(Self::Nav1RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Nav2RadioStbySet => Ok(Self::Nav2RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Nav3RadioStbySet => Ok(Self::Nav3RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Nav4RadioStbySet => Ok(Self::Nav4RadioStbySet {
+                value: (event.dwData as f32 / 1000000.0),
+            }),
+            ClientEventRequest::Nav1RadioSwap => Ok(Self::Nav1RadioSwap),
+            ClientEventRequest::Nav2RadioSwap => Ok(Self::Nav2RadioSwap),
+            ClientEventRequest::Nav3RadioSwap => Ok(Self::Nav3RadioSwap),
+            ClientEventRequest::Nav4RadioSwap => Ok(Self::Nav4RadioSwap),
+            // Transponder
+            ClientEventRequest::TransponderSet => Ok(Self::TransponderSet {
+                value: bcd16_to_u16(event.dwData as u16),
+            }),
         }
     }
 }
 
-impl From<ClientEvent> for (ClientEventRequest, i32) {
+impl From<ClientEvent> for (ClientEventRequest, u32) {
     fn from(event: ClientEvent) -> Self {
         match event {
             // Aircraft Engine
-            ClientEvent::Throttle1Set { value } => (ClientEventRequest::Throttle1Set, value),
-            ClientEvent::Throttle2Set { value } => (ClientEventRequest::Throttle2Set, value),
-            ClientEvent::Throttle3Set { value } => (ClientEventRequest::Throttle3Set, value),
-            ClientEvent::Throttle4Set { value } => (ClientEventRequest::Throttle4Set, value),
+            ClientEvent::Throttle1Set { value } => (ClientEventRequest::Throttle1Set, value as u32),
+            ClientEvent::Throttle2Set { value } => (ClientEventRequest::Throttle2Set, value as u32),
+            ClientEvent::Throttle3Set { value } => (ClientEventRequest::Throttle3Set, value as u32),
+            ClientEvent::Throttle4Set { value } => (ClientEventRequest::Throttle4Set, value as u32),
             // Aircraft Flight Controls
-            ClientEvent::AxisElevatorSet { value } => (ClientEventRequest::AxisElevatorSet, value),
+            ClientEvent::AxisElevatorSet { value } => (ClientEventRequest::AxisElevatorSet, value as u32),
             // Aircraft Miscellaneous Systems
             ClientEvent::Brakes => (ClientEventRequest::Brakes, 0),
             ClientEvent::BrakesLeft => (ClientEventRequest::BrakesLeft, 0),
             ClientEvent::BrakesRight => (ClientEventRequest::BrakesRight, 0),
-            ClientEvent::AxisLeftBrakeSet { value } => {
-                (ClientEventRequest::AxisLeftBrakeSet, value)
-            }
-            ClientEvent::AxisRightBrakeSet { value } => {
-                (ClientEventRequest::AxisRightBrakeSet, value)
-            }
+            ClientEvent::AxisLeftBrakeSet { value } => (
+                ClientEventRequest::AxisLeftBrakeSet,
+                value as u32,
+            ),
+            ClientEvent::AxisRightBrakeSet { value } => (
+                ClientEventRequest::AxisRightBrakeSet,
+                value as u32,
+            ),
             ClientEvent::ParkingBrakes => (ClientEventRequest::ParkingBrakes, 0),
+            // Comm Radios
+            ClientEvent::Com1RadioStbySet { value } => (
+                ClientEventRequest::Com1RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Com2RadioStbySet { value } => (
+                ClientEventRequest::Com2RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Com3RadioStbySet { value } => (
+                ClientEventRequest::Com3RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Com1RadioSwap => (ClientEventRequest::Com1RadioSwap, 0),
+            ClientEvent::Com2RadioSwap => (ClientEventRequest::Com2RadioSwap, 0),
+            ClientEvent::Com3RadioSwap => (ClientEventRequest::Com3RadioSwap, 0),
+            // Nav Radios
+            ClientEvent::Nav1RadioStbySet { value } => (
+                ClientEventRequest::Nav1RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Nav2RadioStbySet { value } => (
+                ClientEventRequest::Nav2RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Nav3RadioStbySet { value } => (
+                ClientEventRequest::Nav3RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Nav4RadioStbySet { value } => (
+                ClientEventRequest::Nav4RadioStbySet,
+                (value * 1000000.0) as u32,
+            ),
+            ClientEvent::Nav1RadioSwap => (ClientEventRequest::Nav1RadioSwap, 0),
+            ClientEvent::Nav2RadioSwap => (ClientEventRequest::Nav2RadioSwap, 0),
+            ClientEvent::Nav3RadioSwap => (ClientEventRequest::Nav3RadioSwap, 0),
+            ClientEvent::Nav4RadioSwap => (ClientEventRequest::Nav4RadioSwap, 0),
+            // Transponder
+            ClientEvent::TransponderSet { value } => (
+                ClientEventRequest::TransponderSet,
+                u16_to_bcd16(value) as u32,
+            )
         }
     }
+}
+
+/// SimConnect Event Flags for TransmitClientEvent
+///
+/// https://www.prepar3d.com/SDKv5/sdk/simconnect_api/references/general_functions.html#SimConnect_TransmitClientEvent
+#[derive(Debug, Copy, Clone, PartialEq, Eq, num_enum::TryFromPrimitive)]
+#[repr(u32)]
+pub enum EventFlag {
+    DoNothing,
+    SlowRepeatTimer,
+    FastRepeatTimer,
+    // GroupIdIsPriority Not currently implemented in SDK
 }